+++ R/plotly.R
@@ -374,7 +374,7 @@ as_widget <- function(x, ...) {
 typedArrayPolyfill <- function() {
   htmltools::htmlDependency(
     "typedarray", 0.1,
-    src = system.file("htmlwidgets", "lib", "typedarray", package = "plotly"),
+    src = depPath("typedarray"),
     script = "typedarray.min.js"
   )
 }
@@ -388,7 +388,7 @@ typedArrayPolyfill <- function() {
 #' @details The polyfill seems to be only relevant for those rendering plots 
 #' via phantomjs and RStudio on some Windows platforms.
 #'
-#' @param a plotly object
+#' @param p a plotly object
 #' @export
 #' @examples 
 #'